pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-k8s_cost_optimizer"
version = "0.1.0"
description = "A production-grade RL environment for learning proactive Kubernetes autoscaling strategies using LLMs."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "fastapi>=0.135.3",
    "gradio>=6.11.0",
    "openai>=2.30.0",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "uvicorn>=0.42.0",
]

[project.optional-dependencies]
dev = [
    "black>=26.3.1",
    "mypy>=1.20.0",
    "pytest>=9.0.2",
    "pytest-cov>=7.1.0",
]

[project.scripts]
server = "k8s_cost_optimizer.server.app:main"
kubecost-gym = "k8s_cost_optimizer.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["k8s_cost_optimizer", "k8s_cost_optimizer.server"]
package-dir = { "k8s_cost_optimizer" = ".", "k8s_cost_optimizer.server" = "server" }
